refactor: split mocks
