+++ packages/account-api/src/mocks/index.ts
@@ -1,3 +1,5 @@
 export * from './accounts';
+export * from './keyrings';
 export * from './groups';
 export * from './wallets';
+export * from './snaps';
