+++ packages/account-api/src/mocks/keyrings.ts
@@ -1,0 +1,4 @@
+export const MOCK_ENTROPY_SOURCE_1 = 'mock-keyring-id-1';
+export const MOCK_ENTROPY_SOURCE_2 = 'mock-keyring-id-2';
+
+export const MOCK_PRIVATE_KEY_KEYRING_TYPE = 'Simple Key Pair';
