+++ packages/account-api/src/mocks/snaps.ts
@@ -1,0 +1,17 @@
+export const MOCK_SNAP_1 = {
+  id: 'local:mock-snap-id-1',
+  name: 'Mock Snap 1',
+  enabled: true,
+  manifest: {
+    proposedName: 'Mock Snap 1',
+  },
+};
+
+export const MOCK_SNAP_2 = {
+  id: 'npm:@metamask/mock-snap-id-2',
+  name: 'Mock Snap 2',
+  enabled: true,
+  manifest: {
+    proposedName: 'Mock Snap 2',
+  },
+};
