+++ packages/account-api/src/mocks/wallets.ts
@@ -1,9 +1,8 @@
 import {
   MOCK_ENTROPY_SOURCE_1,
   MOCK_PRIVATE_KEY_KEYRING_TYPE,
-  MOCK_SNAP_1,
-  MOCK_SNAP_2,
-} from './accounts';
+} from './keyrings';
+import { MOCK_SNAP_1, MOCK_SNAP_2 } from './snaps';
 import { AccountWalletType } from '../api';
 
 export const MOCK_ENTROPY_WALLET_ID = `${AccountWalletType.Entropy}:${MOCK_ENTROPY_SOURCE_1}`;
